requirements.txt
click
loguru
fastapi
uvicorn
pydantic
watchfiles
python-dotenv
jieba
